# feat(telemetry): implement performance testing framework (Phase 8)