… internal/telemetry/integration_test.go
… internal/telemetry/optimized_benchmark_test.go