## Summary

This PR implements Phase 8 of the telemetry system migration (#833), focusing on comprehensive performance testing and validation. The primary goal was to ensure the telemetry system has minimal performance impact when disabled (<100ns) while providing robust testing capabilities.

## Key Achievements

### 🎯 Performance Goals Met
- **2.4 nanoseconds** per operation when telemetry is disabled (target: <100ns)
- **Zero memory allocations** on the disabled path
- Atomic flag checking optimized to 1.3ns

### 🧪 Testing Infrastructure
- **MockTransport**: Thread-safe Sentry transport implementation for testing
- **Test Helpers**: Unified testing interface for both `testing.T` and `testing.B`
- **Integration Tests**: Complete end-to-end telemetry flow validation
- **Performance Benchmarks**: Comprehensive benchmark suite

## What's Changed

### MockTransport Implementation
- Implements full `sentry.Transport` interface
- Thread-safe event capture and retrieval
- Helper methods for test assertions
- Support for async event verification

### Test Coverage
- ✅ Telemetry system unit tests
- ✅ Integration tests with error package
- ✅ End-to-end flow tests
- ✅ Privacy compliance verification
- ✅ Concurrent operation tests
- ✅ Performance benchmarks

### Performance Optimizations
- Atomic flag for fast telemetry state checking
- Optimized capture functions with early returns
- Zero-allocation path when disabled

## Performance Results

```
BenchmarkOptimizedTelemetryDisabled/FastCaptureError-4     496724498    2.423 ns/op    0 B/op    0 allocs/op
BenchmarkOptimizedTelemetryDisabled/FastCaptureMessage-4   491951907    2.448 ns/op    0 B/op    0 allocs/op
BenchmarkOptimizedTelemetryDisabled/AtomicCheck-4          897079670    1.346 ns/op    0 B/op    0 allocs/op
```

## Testing Guidelines

### Using MockTransport
```go
config, cleanup := telemetry.InitForTesting(t)
defer cleanup()

// Your test code here
telemetry.CaptureError(err, "component")

// Verify
telemetry.AssertEventCount(t, config.MockTransport, 1, 100*time.Millisecond)
```

### Performance Testing
```go
// Use optimized functions in production code
if telemetry.IsTelemetryEnabled() {
    telemetry.CaptureError(err, component)
}
```

## Files Changed
- `internal/telemetry/mock_transport.go` - MockTransport implementation
- `internal/telemetry/test_helpers.go` - Testing utilities
- `internal/telemetry/integration_test.go` - Integration tests
- `internal/telemetry/e2e_test.go` - End-to-end tests
- `internal/telemetry/benchmark_test.go` - Performance benchmarks
- `internal/telemetry/optimized_capture.go` - Performance optimizations
- `internal/telemetry/optimized_benchmark_test.go` - Optimized benchmarks

## Related Issues
- Implements Phase 8 of #833
- Continues work from PR #839 (Phase 7)

## Checklist
- [x] Tests pass
- [x] Linter passes
- [x] Performance targets met
- [x] Documentation updated
- [x] No breaking changes

## Next Steps
Phase 9 will focus on documentation and examples to help developers integrate with the new telemetry system.

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>